Commit 9 : Hasil Akhir Masih Error
- Sudah mengikuti semua yang diperintahkan oleh Cursor Namun akhirnya saat website dijalankan tetap error/tidak ditemukan

+++ core/Autoloader.php
@@ -5,11 +5,15 @@ class Autoloader {
     public static function register() {
         spl_autoload_register(function ($class) {
             // Convert namespace to full file path
-            $file = str_replace('\\', DIRECTORY_SEPARATOR, $class) . '.php';
-            $file = __DIR__ . '/../' . $file;
+            $file = $class;
+            $file = str_replace('App\\Controllers\\', 'app/controllers/', $file);
+            $file = str_replace('App\\Models\\', 'app/Models/', $file);
+            $file = str_replace('Core\\', 'core/', $file);
+            $file = str_replace(['\\', '/'], DIRECTORY_SEPARATOR, $file) . '.php';
+            $file = realpath(__DIR__ . '/../' . $file);
             
             // If the file exists, require it
-            if (file_exists($file)) {
+            if ($file && file_exists($file)) {
                 echo "<pre>Autoloading: $file</pre>";
                 require $file;
                 return true;
